Pom cleanup, worn-out tyre and purchase order view added
