+++ src/main/java/com/vaibhav/tyremanager/repository/ITyreInventoryRepository.java
@@ -1,5 +1,7 @@
 package com.vaibhav.tyremanager.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,5 +9,5 @@ import com.vaibhav.tyremanager.models.TyreInventory;
 
 @Repository
 public interface ITyreInventoryRepository extends JpaRepository<TyreInventory, Integer> {
-
+	List<TyreInventory> findByThresholdFlag(String flag);
 }
