+++ src/main/java/com/vaibhav/tyremanager/service/TyreService.java
@@ -7,5 +7,6 @@ import com.vaibhav.tyremanager.models.TyreInventory;
 public interface TyreService {
 	
 	List<TyreInventory> getAllTyres();
+	List<TyreInventory> wornOutTyres();
 	
 }
