+++ src/main/resources/templates/order_book.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-	<title>Home</title>
+	<title>Purchase Orders</title>
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 <!--===============================================================================================-->
@@ -28,65 +28,49 @@
 	<link rel="stylesheet" type="text/css"  th:href="@{/css/util.css}">
 	<link rel="stylesheet" type="text/css" th:href="@{/css/main.css}">
 <!--===============================================================================================-->
+<link data-th-href="@{/webjars/bulma/0.6.0/css/bulma.css}"
+	rel="stylesheet">
+<link rel="stylesheet"
+	data-th-href="@{/webjars/font-awesome/4.7.0/css/font-awesome.min.css}" />
 </head>
 <body>
 
 
 	<div class="container-contact100">
 		<div class="wrap-contact100">
-			<form class="contact100-form validate-form">
+		<form class="contact100-form" action="#" th:action="@{/home}" method="get">
 				<span class="contact100-form-title">
-					Tyre Management Home
+					Purchase Orders
 				</span>
-				<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/addtrip">
+				<table class="table input100 is-hoverable">
+			<tr>
+				<th>Order ID</th>
+				<th>Brand Name</th>
+				<th>Manufacturer</th>
+				<th>Quantity</th>
+				<th>Price</th>
+				<th>Order Date</th>
+			</tr>
+			<tbody>
+				<tr data-th-each="order : ${orders}">
+					<td>[[${order.id}]]</td>
+					<td>[[${order.brand.brandName}]]</td>
+					<td>[[${order.brand.tyreManufacturer.manufacturerName}]]</td>
+					<td>[[${order.quantity}]]</td>
+					<td>[[${order.price}]]</td>
+					<td>[[${order.orderDate}]]</td>
+				</tr>
+			</tbody>
+		</table>
+						<span class="contact100-form-title">
+				</span>
+				<div class="container-contact100-form-btn">
+					<button class="contact100-form-btn">
 						<span>
-							Add Trip
+							Home
+							<i class="fa fa-long-arrow-right m-l-7" aria-hidden="true"></i>
 						</span>
-					</a>
-				</div>
-				<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/newpurchaseorder">
-						<span>
-							New Order
-						</span>
-					</a>
-				</div>
-				<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/tripdashboard">
-						<span>
-							Trip Dashboard
-						</span>
-					</a>
-				</div>
-					<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/tyredashboard">
-						<span>
-							Tyre Dashboard
-						</span>
-					</a>
-				</div>
-
-					<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/providerdashboard">
-						<span>
-							Bus Provider Dashboard
-						</span>
-					</a>
-				</div>
-				<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/purchaseorders">
-						<span>
-							Purchase Orders
-						</span>
-					</a>
-				</div>
-				<div class="wrap-input100 bg1 rs1-wrap-input100">
-					<a class="contact100-form-btn" href="/wornout">
-						<span>
-							Worn-Out Tyres
-						</span>
-					</a>
+					</button>
 				</div>
 			</form>
 		</div>
