+++ src/main/resources/templates/provider_dash.html
@@ -54,7 +54,7 @@
 					<td>[[${provider.providerId}]]</td>
 					<td>[[${provider.providerName}]]</td>
 					<td>[[${provider.description}]]</td>
-					<td><code>[[${provider.buses.size()}]]</code></td>
+					<td>[[${provider.buses.size()}]]</td>
 				</tr>
 			</tbody>
 		</table>
