+++ src/main/resources/templates/trip_dash.html
@@ -61,7 +61,7 @@
 					<td>[[${trip.bus.busProvider.providerName}]]</td>
 					<td><time data-th-attr="date=${trip.tripDate}">[[${trip.tripDate}]]</time>
 					</td>
-					<td><code>[[${trip.distance}]] km</code></td>
+					<td>[[${trip.distance}]] km</td>
 				</tr>
 			</tbody>
 		</table>
